Add new structure for data loading with adding tests

+++ denspp/offline/csv_handler.py
@@ -65,7 +65,7 @@ class CsvHandler:
 
         np.savetxt(self.__path2chck, data, **cmds)
 
-    def read_data_from_csv(self, include_chapter_line: bool = False, start_line: int=0, type_load=float) -> np.array:
+    def read_data_from_csv(self, include_chapter_line: bool = False, start_line: int=0, type_load=int) -> np.array:
         """Reading data in numpy format from csv file
         :param include_chapter_line:    Boolean for including the chapter line
         :param start_line:              Number of rows to skip (exclude chapter line)
